Create Employee gross_pay
Develop a program that will determine the gross pay for each of several employees.
The company pays “straight time” for the first 40 hours worked by each employee and
pays “time and-a-half” for all hours worked in excess of 40 hours. Take input from user
for three employees of the company, the number of hours each employee worked last
week and the hourly rate of each employee. Your program should input this information
for each employee and should determine and display the employee's gross pay.